//ui/views/metadata moved to //ui/base/metadata.
Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/a66b3eed4ce5dbd08d9b71f844b6194fe0a50e3c

commit a66b3eed4ce5dbd08d9b71f844b6194fe0a50e3c
Author: Allen Bauer <kylixrd@chromium.org>
Date:   Wed May 5 02:42:12 2021 +0000

    Moved core ui/views/metadata to ui/base/metadata

    Bug: 1203897
